requirements.txt
Flask==2.3.3
openai==1.3.0
google-generativeai==0.3.2
Pillow>=10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7 